requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
fake-useragent==2.2.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pydantic==2.12.3
pydantic_core==2.41.4
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
psutil==7.1.1
torch==2.3.1+cu121
torchvision==0.18.1+cu121
paddlepaddle-gpu==2.6.1
paddleocr==2.7.1.3
opencv-python==4.9.0.80
pdf2image==1.17.0
Pillow==10.4.0
numpy==1.26.4
scipy==1.13.1
pyclipper==1.3.0.post5
shapely==2.0.4
lmdb==1.4.1
tqdm==4.66.4
nltk==3.8.1
visualdl==2.5.5
python-Levenshtein==0.23.0
python-multipart==0.0.9